volcano options refactor; plot selection variant
