+++ src/components/Enrichment/EnrichmentSVGPlot.scss
@@ -24,14 +24,14 @@
     min-height: 75vh !important;
   }
 
-  .DiffTableIcon {
-    display: none !important;
-    position: absolute !important;
-    top: 10px !important;
-    // left: -2px !important;
-    cursor: pointer !important;
-    color: #1678c2 !important;
-  }
+  // .DiffTableIcon {
+  //   display: none !important;
+  //   position: absolute !important;
+  //   top: 10px !important;
+  //   // left: -2px !important;
+  //   cursor: pointer !important;
+  //   color: #1678c2 !important;
+  // }
 
   .NoPlotsMessage {
     margin: 40px auto !important;
