+++ src/components/Shared/SVGPlot.scss
@@ -10,6 +10,7 @@
     z-index: 5 !important;
     margin-top: 0px;
     margin-bottom: 10px !important;
+    margin-left: 40px !important;
   }
 
   .ui.secondary.pointing.menu {
@@ -26,12 +27,12 @@
   }
 
   .DiffTableIcon {
-    display: none !important;
+    // display: none !important;
     position: absolute !important;
-    top: 10px !important;
-    // left: -2px !important;
+    top: 0px !important;
+    left: 10px !important;
     cursor: pointer !important;
-    color: #1678c2 !important;
+    // color: #1678c2 !important;
   }
 
   .NoPlotsMessage {
@@ -70,4 +71,9 @@
   max-width: fit-content !important;
   z-index: 5 !important;
   margin: 10px 10px 20px 10px !important;
+}
+
+#svgPlotDropdownDifferential {
+  left: 90px !important;
+  top: 0px;
 }
